requirements.txt
streamlit
openai
faiss-cpu
numpy
nltk
python-dotenv
youtube-transcript-api